+++ frameworks/CorePlot/Source/NSCoderExtensions.m
@@ -142,7 +142,7 @@ void MyCGPathApplierFunc(void *info, const CGPathElement *element)
 		NSDictionary *elementData = [pathData objectAtIndex:i];
 		
 		CGPathElementType type = [[elementData objectForKey:@"type"] intValue];
-		newKey = [[NSString alloc] initWithFormat:@"%@[%lu].type", key, i];
+		newKey = [[NSString alloc] initWithFormat:@"%@[%u].type", key, (unsigned int)i];
 		[self encodeInteger:type forKey:newKey];
 		[newKey release];
 		
@@ -152,14 +152,14 @@ void MyCGPathApplierFunc(void *info, const CGPathElement *element)
 			case kCGPathElementAddCurveToPoint: // 3 points
 				point.x = [[elementData objectForKey:@"point3.x"] cgFloatValue];
 				point.y = [[elementData objectForKey:@"point3.y"] cgFloatValue];
-				newKey = [[NSString alloc] initWithFormat:@"%@[%lu].point3", key, i];
+				newKey = [[NSString alloc] initWithFormat:@"%@[%u].point3", key, (unsigned int)i];
 				[self encodeCPTPoint:point forKey:newKey];
 				[newKey release];
 				
 			case kCGPathElementAddQuadCurveToPoint: // 2 points
 				point.x = [[elementData objectForKey:@"point2.x"] cgFloatValue];
 				point.y = [[elementData objectForKey:@"point2.y"] cgFloatValue];
-				newKey = [[NSString alloc] initWithFormat:@"%@[%lu].point2", key, i];
+				newKey = [[NSString alloc] initWithFormat:@"%@[%u].point2", key, (unsigned int)i];
 				[self encodeCPTPoint:point forKey:newKey];
 				[newKey release];
 				
@@ -167,7 +167,7 @@ void MyCGPathApplierFunc(void *info, const CGPathElement *element)
 			case kCGPathElementAddLineToPoint: // 1 point
 				point.x = [[elementData objectForKey:@"point1.x"] cgFloatValue];
 				point.y = [[elementData objectForKey:@"point1.y"] cgFloatValue];
-				newKey = [[NSString alloc] initWithFormat:@"%@[%lu].point1", key, i];
+				newKey = [[NSString alloc] initWithFormat:@"%@[%u].point1", key, (unsigned int)i];
 				[self encodeCPTPoint:point forKey:newKey];
 				[newKey release];
 				break;
@@ -374,7 +374,7 @@ void MyCGPathApplierFunc(void *info, const CGPathElement *element)
 	
 	// decode elements
 	for ( NSUInteger i = 0; i < count; i++ ) {
-		newKey = [[NSString alloc] initWithFormat:@"%@[%lu].type", key, i];
+		newKey = [[NSString alloc] initWithFormat:@"%@[%u].type", key, (unsigned int)i];
 		CGPathElementType type = [self decodeIntegerForKey:newKey];
 		[newKey release];
 		
@@ -382,18 +382,18 @@ void MyCGPathApplierFunc(void *info, const CGPathElement *element)
 		
 		switch ( type ) {
 			case kCGPathElementAddCurveToPoint: // 3 points
-				newKey = [[NSString alloc] initWithFormat:@"%@[%lu].point3", key, i];
+				newKey = [[NSString alloc] initWithFormat:@"%@[%u].point3", key, (unsigned int)i];
 				point3 = [self decodeCPTPointForKey:newKey];
 				[newKey release];
 				
 			case kCGPathElementAddQuadCurveToPoint: // 2 points
-				newKey = [[NSString alloc] initWithFormat:@"%@[%lu].point2", key, i];
+				newKey = [[NSString alloc] initWithFormat:@"%@[%u].point2", key, (unsigned int)i];
 				point2 = [self decodeCPTPointForKey:newKey];
 				[newKey release];
 				
 			case kCGPathElementMoveToPoint: // 1 point
 			case kCGPathElementAddLineToPoint: // 1 point
-				newKey = [[NSString alloc] initWithFormat:@"%@[%lu].point1", key, i];
+				newKey = [[NSString alloc] initWithFormat:@"%@[%u].point1", key, (unsigned int)i];
 				point1 = [self decodeCPTPointForKey:newKey];
 				[newKey release];
 				break;
